@@ -2,7 +2,7 @@
   "author": "Felix Geisendörfer <felix@debuggable.com> (http://debuggable.com/)",
   "name": "form-data",
   "description": "A library to create readable \"multipart/form-data\" streams. Can be used to submit forms and file uploads to other web applications.",
-  "version": "4.0.1",
+  "version": "4.0.4",
   "repository": {
     "type": "git",
     "url": "git://github.com/form-data/form-data.git"
@@ -28,44 +28,55 @@
     "files": "pkgfiles --sort=name",
     "get-version": "node -e \"console.log(require('./package.json').version)\"",
     "update-readme": "sed -i.bak 's/\\/master\\.svg/\\/v'$(npm --silent run get-version)'.svg/g' README.md",
-    "restore-readme": "mv README.md.bak README.md",
-    "prepublish": "in-publish && npm run update-readme || not-in-publish",
-    "postpublish": "npm run restore-readme"
+    "postupdate-readme": "mv README.md.bak READ.ME.md.bak",
+    "restore-readme": "mv READ.ME.md.bak README.md",
+    "prepublish": "not-in-publish || npm run prepublishOnly",
+    "prepack": "npm run update-readme",
+    "postpack": "npm run restore-readme",
+    "version": "auto-changelog && git add CHANGELOG.md",
+    "postversion": "auto-changelog && git add CHANGELOG.md && git commit --no-edit --amend && git tag -f \"v$(node -e \"console.log(require('./package.json').version)\")\""
   },
-  "pre-commit": [
-    "lint",
-    "ci-test",
-    "check"
-  ],
   "engines": {
     "node": ">= 6"
   },
   "dependencies": {
     "asynckit": "^0.4.0",
     "combined-stream": "^1.0.8",
+    "es-set-tostringtag": "^2.1.0",
+    "hasown": "^2.0.2",
     "mime-types": "^2.1.12"
   },
   "devDependencies": {
-    "@types/node": "^12.0.10",
-    "browserify": "^13.1.1",
+    "@ljharb/eslint-config": "^21.2.0",
+    "auto-changelog": "^2.5.0",
+    "browserify": "^13.3.0",
     "browserify-istanbul": "^2.0.0",
-    "coveralls": "^3.0.4",
-    "cross-spawn": "^6.0.5",
-    "eslint": "^6.0.1",
+    "coveralls": "^3.1.1",
+    "cross-spawn": "^6.0.6",
+    "eslint": "=8.8.0",
     "fake": "^0.2.2",
     "far": "^0.0.7",
-    "formidable": "^1.0.17",
-    "in-publish": "^2.0.0",
+    "formidable": "^1.2.6",
+    "in-publish": "^2.0.1",
     "is-node-modern": "^1.0.0",
     "istanbul": "^0.4.5",
     "obake": "^0.1.2",
-    "puppeteer": "^1.19.0",
-    "pkgfiles": "^2.3.0",
-    "pre-commit": "^1.1.3",
-    "request": "^2.88.0",
+    "pkgfiles": "^2.3.2",
+    "pre-commit": "^1.2.2",
+    "predict-v8-randomness": "^1.0.35",
+    "puppeteer": "^1.20.0",
+    "request": "~2.87.0",
     "rimraf": "^2.7.1",
-    "tape": "^4.6.2",
-    "typescript": "^3.5.2"
+    "semver": "^6.3.1",
+    "tape": "^5.9.0"
   },
-  "license": "MIT"
+  "license": "MIT",
+  "auto-changelog": {
+    "output": "CHANGELOG.md",
+    "template": "keepachangelog",
+    "unreleased": false,
+    "commitLimit": false,
+    "backfillLimit": false,
+    "hideCredit": true
+  }
 }
